- refactor(): Modularized the main app crate
