+++ src/rushx_core/error.rs
@@ -1,0 +1,1 @@
+// Shared error types will live here.
