+++ src/rushx_core/state.rs
@@ -1,0 +1,1 @@
+// Shared shell state will live here.
