+++ src/rushx_expand/path.rs
@@ -1,0 +1,1 @@
+// PATH resolution and search helpers will live here.
