+++ src/rushx_expand/vars.rs
@@ -1,0 +1,1 @@
+// Variable expansion logic will live here.
